refactor: delete old parser
