+++ src/utils/string_utils.rs
@@ -1,56 +1,5 @@
 use std::{iter::Peekable, str::Chars};
 
-/// Find all occurrences of a character outside of quotes in a string.
-/// Ej: find_all_outside_quotes(":':':", ':') -> [0, 4]
-///
-/// Escaped values are ignored
-pub fn find_all_outside_quotes<T: AsRef<str>>(str: T, needle: char) -> Vec<usize> {
-    let mut indexes = Vec::new();
-
-    let mut quoted = false;
-    let mut escaped = false;
-    for (i, c) in str.as_ref().chars().enumerate() {
-        match (escaped, quoted, c) {
-            (_, _, '\\') => escaped = !escaped,
-            (false, false, '"') => quoted = true,
-            (false, false, _) => {
-                if c == needle {
-                    indexes.push(i);
-                }
-            }
-            (false, true, '"') => quoted = false,
-            (_, _, _) => {}
-        }
-    }
-
-    indexes
-}
-
-pub fn find_all_outside_quotes_and_brackets<T: AsRef<str>>(str: T, needle: char) -> Vec<usize> {
-    let mut indexes = Vec::new();
-
-    let mut quoted = false;
-    let mut escaped = false;
-    let mut bracket_count = 0;
-    for (i, c) in str.as_ref().chars().enumerate() {
-        match (escaped, quoted, bracket_count, c) {
-            (_, _, _, '\\') => escaped = !escaped,
-            (false, false, _, '"') => quoted = true,
-            (false, false, _, '[') => bracket_count += 1,
-            (false, false, _, ']') => bracket_count -= 1,
-            (false, false, 0, _) => {
-                if c == needle {
-                    indexes.push(i);
-                }
-            }
-            (false, true, _, '"') => quoted = false,
-            (_, _, _, _) => {}
-        }
-    }
-
-    indexes
-}
-
 pub fn wildcard_match_internal(mut haystack: Chars, mut needle: Peekable<Chars>) -> bool {
     loop {
         match needle.peek() {
@@ -89,36 +38,6 @@ pub fn wildcard_match(haystack: &str, needle: &str) -> bool {
 #[cfg(test)]
 mod test {
     use crate::utils::string_utils::wildcard_match;
-
-    #[test]
-    fn test_find_all_outside_quotes() {
-        let needle = ':';
-        let empty = Vec::<usize>::new();
-
-        // Simple case
-        let result = super::find_all_outside_quotes(r#":"::":"#, needle);
-        assert_eq!(vec![0, 5], result);
-
-        // No colons
-        let result = super::find_all_outside_quotes("abcd", needle);
-        assert_eq!(Vec::<usize>::new(), result);
-
-        // All colons inside quotes
-        let result = super::find_all_outside_quotes(r#""a:b:c""#, needle);
-        assert_eq!(Vec::<usize>::new(), result);
-
-        // Colons inside and outside quotes
-        let result = super::find_all_outside_quotes(r#":"a:b":c:d"#, needle);
-        assert_eq!(vec![0, 6, 8], result);
-
-        // Escaped quotes
-        let result = super::find_all_outside_quotes(r#""a\"b:c\":d:e""#, needle);
-        assert_eq!(vec![9, 11], result);
-
-        // Escaped escape
-        let result = super::find_all_outside_quotes(r#"\\":""#, needle);
-        assert_eq!(empty, result);
-    }
 
     #[test]
     fn test_wildcard_match() {
